Tailwind TypeScript config
